+++ packages/ui-core/src/components/Header/doc/Header.docz.tsx
@@ -1,0 +1,7 @@
+export const tableRowStyle = {
+    border: '1px solid #EDEDED',
+    display: 'flex',
+    alignItems: 'center',
+    minHeight: '50px',
+    padding: '15px',
+};