+++ packages/ui-core/src/components/IconButton/IconButton.tsx
@@ -5,29 +5,29 @@ import './IconButton.less';
 import Button from 'components/Button/Button';
 
 export interface IIconButtonProps {
-    /** Theme */
+    /** Цветовая тема */
     theme?: 'green' | 'purple' | 'white' | 'black';
-    /** Type */
+    /** Тип отображения */
     type?: 'primary' | 'outline';
-    /** Link */
+    /** Ссылка */
     href?: string;
-    /** Target - property of <a> tag */
+    /** Target - аргумент тега <a> */
     target?: '_self' | '_blank' | '_parent' | '_top';
-    /** Size for all devices */
+    /** Размер на всех разрешениях экрана */
     sizeAll?: 'small' | 'medium' | 'large';
-    /** Size for wide devices 1280+ */
+    /** Размер на разрешении экрана 1280+ */
     sizeWide?: 'small' | 'medium' | 'large';
-    /** Size for desktop 1020+ */
+    /** Размер на разрешении экрана 1020+ */
     sizeDesktop?: 'small' | 'medium' | 'large';
-    /** Size for tablet 730-1020 */
+    /** Размер на разрешении экрана 730-1020 */
     sizeTablet?: 'small' | 'medium' | 'large';
-    /** Size for mobile 320-730 */
+    /** Размер на разрешении экрана 320-730 */
     sizeMobile?: 'small' | 'medium' | 'large';
-    /** Icon */
+    /** Иконка */
     icon: JSX.Element;
-    /** Disabled */
+    /** Управление возможности взаимодействия с компонентом */
     disabled?: boolean;
-    /** Click event handler */
+    /** Обработчик клика по кнопке */
     onClick?: (e: React.SyntheticEvent<EventTarget>) => void;
 }