+++ packages/ui-core/src/components/IconButton/doc/IconButton.docz.tsx
@@ -1,0 +1,9 @@
+export const blockStyle: React.CSSProperties = {
+    backgroundColor: '#F5F5F5',
+    display: 'inline-grid',
+    gridGap: '20px',
+    gridTemplateColumns: 'repeat(auto-fill, minmax(160px, 1fr))',
+    alignItems: 'center',
+    width: '100%',
+    textAlign: 'center',
+};