+++ packages/ui-core/src/components/Link/Link.docz.tsx
@@ -1,5 +1,0 @@
-import * as React from 'react';
-
-export const LinkWrapper: React.FC = ({ children }) => (
-    <div style={{ padding: 16 }}>{children}</div>
-);