+++ packages/ui-shared/gulpfile.js
@@ -232,7 +232,6 @@ const inlineSvgToReact = () => changePipe(async function (file, encoding) {
 
             try {
                 code = await svgr(data, {
-                    icon: true,
                     svgo: true,
                     svgoConfig: {
                         plugins: [