+++ packages/ui-shared/src/components/VideoBlock/doc/VideoBlock.docz.tsx
@@ -1,0 +1,16 @@
+import { IContent } from '../VideoBlock';
+import video from './video.mp4';
+
+const content: IContent = {
+    title: 'Lorem ipsum',
+    description: [
+        'Lorem ipsum dolor sit amet, consectetuer adipiscing elit. Aenean commodo ligula eget dolor. Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.',
+        'Aenean massa. Cum sociis natoque penatibus et magnis dis parturient montes, nascetur ridiculus mus.',
+    ],
+    href: '#',
+    buttonTitle: 'Lorem ipsum',
+};
+
+const youtubeVideoId = '2Sps5MnvlKM';
+
+export { content, video, youtubeVideoId };