+++ packages/ui-core/CHANGELOG.md
@@ -3,6 +3,73 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.32](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.31...@megafon/ui-core@2.0.0-beta.32) (2020-11-02)
+
+**Note:** Version bump only for package @megafon/ui-core
+
+
+
+
+
+# [2.0.0-beta.31](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.30...@megafon/ui-core@2.0.0-beta.31) (2020-11-02)
+
+
+### Bug Fixes
+
+* **base styles:** edit media queries ([e7a22db](https://github.com/MegafonWebLab/megafon-ui/commit/e7a22dbf98835c163fb77bd1974c8f92d1acf0bd))
+* **grid:** added custom classname prop ([66e9536](https://github.com/MegafonWebLab/megafon-ui/commit/66e953696e74b7648f83b11a72c51ea93d9ff8f4))
+* **icons:** new Cancel svg files with merged path, for all icon sizes ([9d4cf8d](https://github.com/MegafonWebLab/megafon-ui/commit/9d4cf8d85d4c5c24da7b6e7dc159630a4f930034))
+
+
+
+
+
+# [2.0.0-beta.30](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.29...@megafon/ui-core@2.0.0-beta.30) (2020-11-01)
+
+**Note:** Version bump only for package @megafon/ui-core
+
+
+
+
+
+# [2.0.0-beta.29](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.28...@megafon/ui-core@2.0.0-beta.29) (2020-10-30)
+
+**Note:** Version bump only for package @megafon/ui-core
+
+
+
+
+
+# [2.0.0-beta.28](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.27...@megafon/ui-core@2.0.0-beta.28) (2020-10-30)
+
+**Note:** Version bump only for package @megafon/ui-core
+
+
+
+
+
+# [2.0.0-beta.27](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.26...@megafon/ui-core@2.0.0-beta.27) (2020-10-29)
+
+
+### Bug Fixes
+
+* **build:** remove icon property from svgr config ([812bc9f](https://github.com/MegafonWebLab/megafon-ui/commit/812bc9fbbea617edbd56e28b6618654876229924))
+
+
+
+
+
+# [2.0.0-beta.26](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.25...@megafon/ui-core@2.0.0-beta.26) (2020-10-29)
+
+
+### Bug Fixes
+
+* **tile:** edit text decoration on hover ([bb5c694](https://github.com/MegafonWebLab/megafon-ui/commit/bb5c694413564f8ec05f1b1c0d5f625efd410e8d))
+
+
+
+
+
 # [2.0.0-beta.25](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-core@2.0.0-beta.24...@megafon/ui-core@2.0.0-beta.25) (2020-10-26)
 
 **Note:** Version bump only for package @megafon/ui-core