+++ packages/ui-shared/CHANGELOG.md
@@ -3,6 +3,68 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.0.0-beta.33](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.32...@megafon/ui-shared@2.0.0-beta.33) (2020-11-02)
+
+**Note:** Version bump only for package @megafon/ui-shared
+
+
+
+
+
+# [2.0.0-beta.32](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.31...@megafon/ui-shared@2.0.0-beta.32) (2020-11-02)
+
+
+### Features
+
+* **video block:** add a new component ([58e857e](https://github.com/MegafonWebLab/megafon-ui/commit/58e857e6584d65c3c85c48f4493a91d868dc5fcb))
+
+
+
+
+
+# [2.0.0-beta.31](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.30...@megafon/ui-shared@2.0.0-beta.31) (2020-11-01)
+
+**Note:** Version bump only for package @megafon/ui-shared
+
+
+
+
+
+# [2.0.0-beta.30](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.29...@megafon/ui-shared@2.0.0-beta.30) (2020-10-30)
+
+**Note:** Version bump only for package @megafon/ui-shared
+
+
+
+
+
+# [2.0.0-beta.29](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.28...@megafon/ui-shared@2.0.0-beta.29) (2020-10-30)
+
+**Note:** Version bump only for package @megafon/ui-shared
+
+
+
+
+
+# [2.0.0-beta.28](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.27...@megafon/ui-shared@2.0.0-beta.28) (2020-10-29)
+
+
+### Bug Fixes
+
+* **build:** remove icon property from svgr config ([812bc9f](https://github.com/MegafonWebLab/megafon-ui/commit/812bc9fbbea617edbd56e28b6618654876229924))
+
+
+
+
+
+# [2.0.0-beta.27](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.26...@megafon/ui-shared@2.0.0-beta.27) (2020-10-29)
+
+**Note:** Version bump only for package @megafon/ui-shared
+
+
+
+
+
 # [2.0.0-beta.26](https://github.com/MegafonWebLab/megafon-ui/compare/@megafon/ui-shared@2.0.0-beta.25...@megafon/ui-shared@2.0.0-beta.26) (2020-10-26)
 
 **Note:** Version bump only for package @megafon/ui-shared